+++ g2product/g2product.go
@@ -23,9 +23,10 @@ import (
 
 type G2product struct {
 	isTrace                           bool
-	logger                            logging.LoggingInterface
-	observers                         subject.Subject
 	LicenseResult                     string
+	logger                            logging.LoggingInterface
+	observerOrigin                    string
+	observers                         subject.Subject
 	ValidateLicenseFileResult         string
 	ValidateLicenseStringBase64Result string
 	VersionResult                     string
@@ -44,7 +45,7 @@ func (client *G2product) getLogger() logging.LoggingInterface {
 		options := []interface{}{
 			&logging.OptionCallerSkip{Value: 4},
 		}
-		client.logger, err = logging.NewSenzingSdkLogger(ProductId, g2productapi.IdMessages, options...)
+		client.logger, err = logging.NewSenzingSdkLogger(ComponentId, g2productapi.IdMessages, options...)
 		if err != nil {
 			panic(err)
 		}
@@ -83,10 +84,23 @@ func (client *G2product) Destroy(ctx context.Context) error {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8001, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8001, err, details)
 		}()
 	}
 	return err
+}
+
+/*
+The GetObserverOrigin method returns the "origin" value of past Observer messages.
+
+Input
+  - ctx: A context to control lifecycle.
+
+Output
+  - The value sent in the Observer's "origin" key/value pair.
+*/
+func (client *G2product) GetObserverOrigin(ctx context.Context) string {
+	return client.observerOrigin
 }
 
 /*
@@ -107,7 +121,7 @@ func (client *G2product) GetSdkId(ctx context.Context) string {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8007, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8007, err, details)
 		}()
 	}
 	return "mock"
@@ -137,7 +151,7 @@ func (client *G2product) Init(ctx context.Context, moduleName string, iniParams 
 				"moduleName":     moduleName,
 				"verboseLogging": strconv.Itoa(verboseLogging),
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8002, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8002, err, details)
 		}()
 	}
 	return err
@@ -163,7 +177,7 @@ func (client *G2product) License(ctx context.Context) (string, error) {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8003, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8003, err, details)
 		}()
 	}
 	return client.LicenseResult, err
@@ -192,7 +206,7 @@ func (client *G2product) RegisterObserver(ctx context.Context, observer observer
 			details := map[string]string{
 				"observerID": observer.GetObserverId(ctx),
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8008, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8008, err, details)
 		}()
 	}
 	return err
@@ -219,10 +233,21 @@ func (client *G2product) SetLogLevel(ctx context.Context, logLevelName string) e
 			details := map[string]string{
 				"logLevel": logLevelName,
 			}
-			notifier.Notify(ctx, client.observers, ProductId, 8009, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8009, err, details)
 		}()
 	}
 	return err
+}
+
+/*
+The SetObserverOrigin method sets the "origin" value in future Observer messages.
+
+Input
+  - ctx: A context to control lifecycle.
+  - origin: The value sent in the Observer's "origin" key/value pair.
+*/
+func (client *G2product) SetObserverOrigin(ctx context.Context, origin string) {
+	client.observerOrigin = origin
 }
 
 /*
@@ -247,7 +272,7 @@ func (client *G2product) UnregisterObserver(ctx context.Context, observer observ
 		details := map[string]string{
 			"observerID": observer.GetObserverId(ctx),
 		}
-		notifier.Notify(ctx, client.observers, ProductId, 8010, err, details)
+		notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8010, err, details)
 	}
 	err = client.observers.UnregisterObserver(ctx, observer)
 	if !client.observers.HasObservers(ctx) {
@@ -280,7 +305,7 @@ func (client *G2product) ValidateLicenseFile(ctx context.Context, licenseFilePat
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8004, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8004, err, details)
 		}()
 	}
 	return client.ValidateLicenseFileResult, err
@@ -311,7 +336,7 @@ func (client *G2product) ValidateLicenseStringBase64(ctx context.Context, licens
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8005, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8005, err, details)
 		}()
 	}
 	return client.ValidateLicenseStringBase64Result, err
@@ -337,7 +362,7 @@ func (client *G2product) Version(ctx context.Context) (string, error) {
 	if client.observers != nil {
 		go func() {
 			details := map[string]string{}
-			notifier.Notify(ctx, client.observers, ProductId, 8006, err, details)
+			notifier.Notify(ctx, client.observers, client.observerOrigin, ComponentId, 8006, err, details)
 		}()
 	}
 	return client.VersionResult, err