+++ CHANGELOG.md
@@ -10,12 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -
 
+## [0.1.4] - 2023-05-10
+
+### Changed in 0.1.4
+
+- Added `GetObserverOrigin()` and `SetObserverOrigin()` to g2* packages
+- Update dependencies
+  - github.com/senzing/g2-sdk-go v0.6.2
+  - github.com/senzing/go-observing v0.2.2
+
 ## [0.1.3] - 2023-04-21
 
 ### Changed in 0.1.3
-
-- Update dependencies
-  - github.com/senzing/g2-sdk-go v0.6.1
 
 ## [0.1.2] - 2023-04-21